feat: firebase logger 적용
