+++ Projects/Feature/SSCreateEnvelope/Sources/Commons/SSLog/CreateType+.swift
@@ -42,42 +42,100 @@ func convertMarketingModuleName(_ createType: CreateType, viewType: CreateEnvelo
   }
 }
 
+func pushButtonLogEvent(_ createType: CreateEnvelopeInitialType, lastPathState state: CreateEnvelopePath.State?) {
+  let createType: CreateType = switch createType {
+  case .sent,
+       .sentWithFriendID:
+    .sent
+  case .received:
+    .received
+  }
+  switch createType {
+  case .sent:
+    ssLogEvent(SentEvents.tappedNextButtonAtCreateEnvelope(type: getViewType(state)))
+  case .received:
+    break
+  }
+}
+
+func backButtonLogEvent(_ createType: CreateEnvelopeInitialType, lastPathState state: CreateEnvelopePath.State?) {
+  let createType: CreateType = switch createType {
+  case .sent,
+       .sentWithFriendID:
+    .sent
+  case .received:
+    .received
+  }
+  switch createType {
+  case .sent:
+    ssLogEvent(SentEvents.tappedBackButtonAtCreateEnvelope(type: getViewType(state)))
+  case .received:
+    break
+  }
+}
+
 func ssLogEvent(
   _ createType: CreateType,
   eventName: String = "",
   lastPathState state: CreateEnvelopePath.State?,
   eventType: SSLogEventType
 ) {
-  let marketingModuleName = convertMarketingModuleName(createType, viewType: .getViewType(lastPathState: state))
+  let marketingModuleName = convertMarketingModuleName(createType, viewType: getViewType(state))
   SSFirebase.ssLogEvent(marketingModuleName, eventName: eventName, eventType: eventType)
 }
 
-extension CreateEnvelopeMarketingModule {
-  static func getViewType(lastPathState: CreateEnvelopePath.State?) -> Self {
-    switch lastPathState {
-    case .createEnvelopePrice:
-      .price
-    case .createEnvelopeName:
-      .name
-    case .createEnvelopeRelation:
-      .relation
-    case .createEnvelopeEvent:
-      .category
-    case .createEnvelopeDate:
-      .date
-    case .createEnvelopeAdditionalSection:
-      .additionalSection
-    case .createEnvelopeAdditionalMemo:
-      .memo
-    case .createEnvelopeAdditionalContact:
-      .contact
-    case .createEnvelopeAdditionalIsGift:
-      .gift
-    case .createEnvelopeAdditionalIsVisitedEvent:
-      .isVisited
-    // 첫 화면에서 요청한 경우 price로 넘어가게 됩니다.
-    case .none:
-      .price
-    }
+private func getViewType(_ lastPathState: CreateEnvelopePath.State?) -> CreateEnvelopeViewTypes {
+  switch lastPathState {
+  case .createEnvelopePrice:
+    .price
+  case .createEnvelopeName:
+    .name
+  case .createEnvelopeRelation:
+    .relation
+  case .createEnvelopeEvent:
+    .category
+  case .createEnvelopeDate:
+    .date
+  case .createEnvelopeAdditionalSection:
+    .selectAdditional
+  case .createEnvelopeAdditionalMemo:
+    .memo
+  case .createEnvelopeAdditionalContact:
+    .phoneNumber
+  case .createEnvelopeAdditionalIsGift:
+    .gift
+  case .createEnvelopeAdditionalIsVisitedEvent:
+    .isVisited
+  // 첫 화면에서 요청한 경우 price로 넘어가게 됩니다.
+  case .none:
+    .price
+  }
+}
+
+private func getViewType(_ lastPathState: CreateEnvelopePath.State?) -> CreateEnvelopeMarketingModule {
+  switch lastPathState {
+  case .createEnvelopePrice:
+    .price
+  case .createEnvelopeName:
+    .name
+  case .createEnvelopeRelation:
+    .relation
+  case .createEnvelopeEvent:
+    .category
+  case .createEnvelopeDate:
+    .date
+  case .createEnvelopeAdditionalSection:
+    .additionalSection
+  case .createEnvelopeAdditionalMemo:
+    .memo
+  case .createEnvelopeAdditionalContact:
+    .contact
+  case .createEnvelopeAdditionalIsGift:
+    .gift
+  case .createEnvelopeAdditionalIsVisitedEvent:
+    .isVisited
+  // 첫 화면에서 요청한 경우 price로 넘어가게 됩니다.
+  case .none:
+    .price
   }
 }
